move emulated firestore to test code
